\name{add_dates}
\alias{add_dates}
\title{Adding date-time objects...}
\usage{add_dates(e1, e2)}

\description{
  Adding date-time objects
}

\details{
  add_dates adds two objects, both of which can be
  date-time objects. Lubridate is built so that you don't
  need to use add_dates but can use regular math
  expressions instead.
}
\arguments{
  \item{e1}{a numeric or date-time object}
  \item{e2}{a numeric or date-time object}
}
