% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.r
\docType{methods}
\name{as_date}
\alias{as_date}
\alias{as_date,POSIXt-method}
\alias{as_date,character-method}
\alias{as_date,numeric-method}
\title{Change an object to a Date}
\usage{
as_date(x, ...)

\S4method{as_date}{POSIXt}(x, tz = NULL)

\S4method{as_date}{numeric}(x, origin = NULL)

\S4method{as_date}{character}(x, format = NULL)
}
\arguments{
\item{x}{a vector of \code{\link{POSIXt}}, numeric or character objects}

\item{...}{further arguments to be passed to specific methods (see above).}

\item{tz}{a time zone name (default: time zone of the POSIXt object
\code{x}). See \code{\link{olson_time_zones}}.}

\item{origin}{a Date object, or something which can be coerced by
\code{as.Date(origin, ...)} to such an object (default: the Unix epoch of
"1970-01-01"). Note that in this instance, \code{x} is
assumed to reflect the number of days since \code{origin} at \code{"UTC"}.}

\item{format}{a character string (default: \code{\%Y-\%m-\%d}). See
\code{\link{strptime}} for alternative format specifications.}
}
\value{
a vector of \code{\link{Date}} objects corresponding to \code{x}.
}
\description{
as_date changes \code{\link{POSIXt}}, numeric and character objects to
\code{\link{Date}} objects.
}

