\name{DateCoercion}
\title{Convert every date method we know about to POSIXlt and POSIXct...}
\usage{as.POSIXlt.fts(x, tz="", ...)}

\description{
  Convert every date method we know about to POSIXlt and POSIXct
}
\keyword{internal}
\alias{as.POSIXlt.fts}
\alias{as.POSIXlt.its}
\alias{as.POSIXlt.timeSeries}
\alias{as.POSIXlt.irts}
\alias{as.POSIXlt.xts}
\alias{as.POSIXlt.zoo}
\alias{as.POSIXlt.tis}
\alias{as.POSIXct.fts}
\alias{as.POSIXct.its}
\alias{as.POSIXct.timeSeries}
\alias{as.POSIXct.irts}
\alias{as.POSIXct.xts}
\alias{as.POSIXct.zoo}
\alias{as.POSIXct.tis}
