% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{Deprecated-lubridate}
\alias{Deprecated-lubridate}
\alias{new_difftime}
\alias{new_duration}
\alias{new_interval}
\alias{new_period}
\title{Deprecated function in \code{lubridate} package}
\usage{
new_period(...)

new_interval(...)

new_duration(...)

new_difftime(...)
}
\arguments{
\item{...}{arguments to be passed to the functions
(obscured to enforce the usage of new functions)}
}
\description{
Deprecated function in \code{lubridate} package
}

