% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r
\name{==,Interval,Interval-method}
\alias{==,Interval,Interval-method}
\title{Equality check for Interval objects}
\usage{
\S4method{==}{Interval,Interval}(e1, e2)
}
\arguments{
\item{e1}{An Interval object}

\item{e2}{Another Interval object}
}
\value{
TRUE if the start and end of \code{e1} match the start and end of \code{e1}, otherwise FALSE
}
\description{
Check whether \code{e1} and \code{e2} have the same start and end date-time objects.
}
\examples{
int1 <- interval(as.Date("2021-01-01"), as.Date("2021-12-30"))
int2 <- interval(as.Date("2021-01-01"), as.Date("2021-12-30"))
int3 <- interval(as.Date("2021-01-02"), as.Date("2021-12-31"))

int1 == int2 # TRUE
int1 == int3 # FALSE
c(int1, int2) == c(int2, int3) # TRUE FALSE
int1 == c(int1, int2, int3) # TRUE TRUE FALSE
c(int1, int2, int3) == int1 # TRUE TRUE FALSE
}
