\name{flip}
\alias{flip}
\title{Shifts an interval up or down the timeline by a specified amount
note this may change the exact length of the interval}
\usage{
  flip(int)
}
\description{
  Shifts an interval up or down the timeline by a specified
  amount note this may change the exact length of the
  interval
}

