% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/round.r
\name{floor_date}
\alias{floor_date}
\title{Round date-times down.}
\usage{
floor_date(x, unit = c("second", "minute", "hour", "day", "week", "month",
  "year", "quarter"))
}
\arguments{
\item{x}{a vector of date-time objects}

\item{unit}{a character string specifying the time unit to be rounded to. Should be one of
"second", "minute", "hour", "day", "week", "month", "quarter", or "year."}
}
\value{
x with the appropriate units floored
}
\description{
\code{floor_date} takes a date-time object and rounds it down to the nearest integer
value of the specified time unit. Users can specify whether to round down to
the nearest second, minute, hour, day, week, month, quarter, or year.
}
\details{
By convention the boundary for a month is the first second of the month. Thus
\code{floor_date(ymd("2000-03-01"), "month")} gives "2000-03-01 UTC".
}
\examples{
x <- as.POSIXct("2009-08-03 12:01:59.23")
floor_date(x, "second")
# "2009-08-03 12:01:59 CDT"
floor_date(x, "minute")
# "2009-08-03 12:01:00 CDT"
floor_date(x, "hour")
# "2009-08-03 12:00:00 CDT"
floor_date(x, "day")
# "2009-08-03 CDT"
floor_date(x, "week")
# "2009-08-02 CDT"
floor_date(x, "month")
# "2009-08-01 CDT"
floor_date(x, "quarter")
# "2009-07-01 CDT"
floor_date(x, "year")
# "2009-01-01 CST"
}
\seealso{
\code{\link{ceiling_date}}, \code{\link{round_date}}
}
\keyword{chron}
\keyword{manip}

