% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/instants.r
\name{here}
\alias{here}
\title{The current time in your local timezone}
\usage{
here()
}
\value{
the current date and time as a POSIXct object
}
\description{
The current time in your local timezone
}
\examples{
here()
}
\seealso{
\code{\link{now}}
}
\keyword{chron}
\keyword{utilities}

