% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.r, R/POSIXt.r, R/intervals.r,
%   R/durations.r, R/periods.r, R/hidden.r
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{day<-,Date-method}
\alias{hour<-,Date-method}
\alias{minute<-,Date-method}
\alias{month<-,Date-method}
\alias{second<-,Date-method}
\alias{year<-,Date-method}
\alias{day<-,POSIXt-method}
\alias{hour<-,POSIXt-method}
\alias{minute<-,POSIXt-method}
\alias{month<-,POSIXt-method}
\alias{second<-,POSIXt-method}
\alias{year<-,POSIXt-method}
\alias{Arith,Interval,ANY-method}
\alias{Arith,ANY,Interval-method}
\alias{intersect,Interval,Interval-method}
\alias{union,Interval,Interval-method}
\alias{setdiff,Interval,Interval-method}
\alias{as.numeric,Interval-method}
\alias{as.integer,Interval-method}
\alias{show,Interval-method}
\alias{c,Interval-method}
\alias{rep,Interval-method}
\alias{[,Interval-method}
\alias{[<-,Interval,ANY,ANY,ANY-method}
\alias{[[,Interval-method}
\alias{[[<-,Interval,ANY,ANY,ANY-method}
\alias{$,Interval-method}
\alias{$<-,Interval-method}
\alias{as.difftime,Interval-method}
\alias{as.character,Interval-method}
\alias{day<-,Interval-method}
\alias{hour<-,Interval-method}
\alias{minute<-,Interval-method}
\alias{month<-,Interval-method}
\alias{second<-,Interval-method}
\alias{year<-,Interval-method}
\alias{+,Interval,Duration-method}
\alias{+,Interval,Interval-method}
\alias{+,Interval,Period-method}
\alias{+,Interval,Date-method}
\alias{+,Date,Interval-method}
\alias{+,Interval,difftime-method}
\alias{+,difftime,Interval-method}
\alias{+,Interval,numeric-method}
\alias{+,numeric,Interval-method}
\alias{+,Interval,POSIXct-method}
\alias{+,POSIXct,Interval-method}
\alias{+,Interval,POSIXlt-method}
\alias{+,POSIXlt,Interval-method}
\alias{/,Interval,Duration-method}
\alias{/,Interval,Interval-method}
\alias{/,Interval,Period-method}
\alias{/,Interval,difftime-method}
\alias{/,difftime,Interval-method}
\alias{/,Interval,numeric-method}
\alias{/,numeric,Interval-method}
\alias{*,Interval,ANY-method}
\alias{*,ANY,Interval-method}
\alias{-,Interval,missing-method}
\alias{-,Interval,Interval-method}
\alias{-,Date,Interval-method}
\alias{-,POSIXct,Interval-method}
\alias{-,POSIXlt,Interval-method}
\alias{-,numeric,Interval-method}
\alias{-,Interval,Date-method}
\alias{-,Interval,POSIXct-method}
\alias{-,Interval,POSIXlt-method}
\alias{-,Interval,numeric-method}
\alias{-,Duration,Interval-method}
\alias{-,Period,Interval-method}
\alias{\%\%,Interval,Duration-method}
\alias{\%\%,Interval,Interval-method}
\alias{\%\%,Interval,Period-method}
\alias{\%\%,Interval,Duration}
\alias{\%\%,Interval,Interval}
\alias{\%\%,Interval,Period}
\alias{-,Date,Interval}
\alias{-,Duration,Interval}
\alias{-,Interval,Date}
\alias{-,Interval,Interval}
\alias{-,Interval,POSIXct}
\alias{-,Interval,POSIXlt}
\alias{-,Interval,numeric}
\alias{-,POSIXct,Interval}
\alias{-,POSIXlt,Interval}
\alias{-,numeric,Interval}
\alias{-,Interval,Duration-method}
\alias{-,Interval,Period-method}
\alias{Compare,Duration,ANY-method}
\alias{Compare,Duration,Duration-method}
\alias{Compare,difftime,Duration-method}
\alias{Compare,ANY,Duration-method}
\alias{Compare,Duration,Period-method}
\alias{Compare,Duration,difftime-method}
\alias{Compare,character,Duration-method}
\alias{Compare,Duration,character-method}
\alias{Compare,numeric,Duration-method}
\alias{Compare,Duration,numeric-method}
\alias{as.numeric,Duration-method}
\alias{as.integer,Duration-method}
\alias{show,Duration-method}
\alias{c,Duration-method}
\alias{rep,Duration-method}
\alias{[,Duration-method}
\alias{[<-,Duration,ANY,ANY,ANY-method}
\alias{[[,Duration-method}
\alias{[[<-,Duration,ANY,ANY,ANY-method}
\alias{$,Duration-method}
\alias{$<-,Duration-method}
\alias{as.difftime,Duration-method}
\alias{day<-,Duration-method}
\alias{hour<-,Duration-method}
\alias{minute<-,Duration-method}
\alias{month<-,Duration-method}
\alias{second<-,Duration-method}
\alias{year<-,Duration-method}
\alias{as.character,Duration-method}
\alias{+,Duration,Duration-method}
\alias{+,Duration,Interval-method}
\alias{+,Duration,Period-method}
\alias{+,Duration,Date-method}
\alias{+,Date,Duration-method}
\alias{+,Duration,difftime-method}
\alias{+,difftime,Duration-method}
\alias{+,Duration,numeric-method}
\alias{+,numeric,Duration-method}
\alias{+,Duration,POSIXct-method}
\alias{+,POSIXct,Duration-method}
\alias{+,Duration,POSIXlt-method}
\alias{+,POSIXlt,Duration-method}
\alias{-,Duration,ANY-method}
\alias{/,Duration,Duration-method}
\alias{/,Duration,Interval-method}
\alias{/,Duration,Period-method}
\alias{/,Duration,difftime-method}
\alias{/,difftime,Duration-method}
\alias{/,Duration,numeric-method}
\alias{/,numeric,Duration-method}
\alias{*,Duration,ANY-method}
\alias{*,ANY,Duration-method}
\alias{\%\%,Duration,Duration-method}
\alias{\%\%,Duration,Interval-method}
\alias{\%\%,Duration,Period-method}
\alias{-,Duration,missing-method}
\alias{-,ANY,Duration-method}
\alias{-,Duration,Duration-method}
\alias{-,Duration,Date-method}
\alias{-,Duration,POSIXct-method}
\alias{-,Duration,POSIXlt-method}
\alias{-,Duration,Period-method}
\alias{-,Duration,numeric-method}
\alias{-,POSIXct,Duration-method}
\alias{-,numeric,Duration-method}
\alias{-,POSIXlt,Duration-method}
\alias{-,Date,Duration-method}
\alias{Arith,ANY,Period-method}
\alias{Arith,Duration,Period-method}
\alias{Arith,Period,Duration-method}
\alias{Compare,Period,Duration-method}
\alias{Compare,numeric,Period-method}
\alias{Compare,difftime,Period-method}
\alias{Compare,Period,difftime-method}
\alias{Compare,Period,Period-method}
\alias{Compare,Period,character-method}
\alias{Compare,Period,numeric-method}
\alias{Compare,character,Period-method}
\alias{second,Period-method}
\alias{second<-,Period-method}
\alias{minute,Period-method}
\alias{minute<-,Period-method}
\alias{hour,Period-method}
\alias{hour<-,Period-method}
\alias{Arith,Period,ANY-method}
\alias{day,Period-method}
\alias{day<-,Period-method}
\alias{month,Period-method}
\alias{month<-,Period-method}
\alias{year,Period-method}
\alias{year<-,Period-method}
\alias{date,Period-method}
\alias{date<-,Period-method}
\alias{as.numeric,Period-method}
\alias{as.integer,Period-method}
\alias{show,Period-method}
\alias{c,Period-method}
\alias{rep,Period-method}
\alias{[,Period-method}
\alias{[<-,Period,ANY,ANY,Period-method}
\alias{[[,Period-method}
\alias{[[<-,Period,ANY,ANY,Period-method}
\alias{$,Period-method}
\alias{$<-,Period-method}
\alias{as.difftime,Period-method}
\alias{as.character,Period-method}
\alias{+,Period,Duration-method}
\alias{+,Period,Interval-method}
\alias{+,Period,Period-method}
\alias{+,Period,Date-method}
\alias{+,Date,Period-method}
\alias{+,Period,difftime-method}
\alias{+,difftime,Period-method}
\alias{+,Period,numeric-method}
\alias{+,numeric,Period-method}
\alias{+,Period,POSIXct-method}
\alias{+,POSIXct,Period-method}
\alias{+,Period,POSIXlt-method}
\alias{+,POSIXlt,Period-method}
\alias{/,Period,Duration-method}
\alias{/,Period,Interval-method}
\alias{/,Period,Period-method}
\alias{/,Period,difftime-method}
\alias{/,difftime,Period-method}
\alias{/,Period,numeric-method}
\alias{/,numeric,Period-method}
\alias{*,Period,ANY-method}
\alias{*,ANY,Period-method}
\alias{-,Period,ANY-method}
\alias{-,Period,missing-method}
\alias{-,ANY,Period-method}
\alias{-,Period,Period-method}
\alias{\%\%,Period,Duration-method}
\alias{\%\%,Period,Interval-method}
\alias{\%\%,Period,Period-method}
\alias{>,Period,Period-method}
\alias{>=,Period,Period-method}
\alias{==,Period,Period-method}
\alias{!=,Period,Period-method}
\alias{<=,Period,Period-method}
\alias{<,Period,Period-method}
\alias{>,Period,Duration-method}
\alias{>=,Period,Duration-method}
\alias{==,Period,Duration-method}
\alias{!=,Period,Duration-method}
\alias{<=,Period,Duration-method}
\alias{<,Period,Duration-method}
\alias{>,Duration,Period-method}
\alias{>=,Duration,Period-method}
\alias{==,Duration,Period-method}
\alias{!=,Duration,Period-method}
\alias{<=,Duration,Period-method}
\alias{<,Duration,Period-method}
\alias{>,Period,numeric-method}
\alias{>=,Period,numeric-method}
\alias{==,Period,numeric-method}
\alias{!=,Period,numeric-method}
\alias{<=,Period,numeric-method}
\alias{<,Period,numeric-method}
\alias{>,numeric,Period-method}
\alias{>=,numeric,Period-method}
\alias{==,numeric,Period-method}
\alias{!=,numeric,Period-method}
\alias{<=,numeric,Period-method}
\alias{<,numeric,Period-method}
\alias{!=,Duration,Period}
\alias{!=,Period,Duration}
\alias{!=,Period,Period}
\alias{!=,Period,numeric}
\alias{!=,numeric,Period}
\alias{\%\%,Period,Duration}
\alias{\%\%,Period,Interval}
\alias{\%\%,Period,Period}
\alias{*,ANY,Period}
\alias{*,Period,ANY}
\alias{-,ANY,Period}
\alias{-,Period,Interval}
\alias{-,Period,missing}
\alias{/,numeric,Period}
\alias{<,Duration,Period}
\alias{<,Period,Duration}
\alias{<,Period,Period}
\alias{<,Period,numeric}
\alias{<,numeric,Period}
\alias{<=,Duration,Period}
\alias{<=,Period,Duration}
\alias{<=,Period,Period}
\alias{<=,Period,numeric}
\alias{<=,numeric,Period}
\alias{==,Duration,Period}
\alias{==,Period,Duration}
\alias{==,Period,Period}
\alias{==,Period,numeric}
\alias{==,numeric,Period}
\alias{>,Duration,Period}
\alias{>,Period,Duration}
\alias{>,Period,Period}
\alias{>,Period,numeric}
\alias{>,numeric,Period}
\alias{>=,Duration,Period}
\alias{>=,Period,Duration}
\alias{>=,Period,Period}
\alias{>=,Period,numeric}
\alias{>=,numeric,Period}
\alias{-,Period,Date-method}
\alias{-,Period,Duration-method}
\alias{-,Period,POSIXct-method}
\alias{-,Period,POSIXlt-method}
\alias{-,Period,numeric-method}
\alias{-,Date,Period-method}
\alias{-,numeric,Period-method}
\alias{-,POSIXlt,Period-method}
\alias{-,POSIXct,Period-method}
\title{Internal page for hidden aliases}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
