\name{hour-set}
\alias{hour-set}
\title{Internal function.}
\usage{\method{hour<-}{default} (x, value)}

\description{
  Internal function. Replaces the hours element of a date with a
  specified value.
}
\seealso{\code{\link{hour}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's hours component.}
}
