\name{int_overlaps}
\alias{int_overlaps}
\title{Test if two intervals overlap}
\usage{
  int_overlaps(int1, int2)
}
\arguments{
  \item{int1}{an Interval object}

  \item{int2}{an Interval object}
}
\value{
  Logical. TRUE if int1 and int2 overlap by at least one
  second. FALSE otherwise.
}
\description{
  Test if two intervals overlap
}

