% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lakers}
\alias{lakers}
\title{Lakers 2008-2009 basketball data set}
\description{
This data set contains play by play statistics of each Los
Angeles Lakers basketball game in the 2008-2009 season. Data
includes the date, opponent, and type of each game (home or
away). Each play is described by the time on the game clock
when the play was made, the period in which the play was
attempted, the type of play, the player and team who made the
play, the result of the play, and the location on the court
where each play was made.
}
\references{
\url{http://www.basketballgeek.com/data/}
}
\keyword{data}

