% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{lubridate-vctrs}
\alias{lubridate-vctrs}
\alias{vec_ptype2.Period}
\alias{vec_ptype2.Duration}
\alias{vec_ptype2.Interval}
\title{vctrs methods for lubridate}
\usage{
vec_ptype2.Period(x, y, ...)

vec_ptype2.Duration(x, y, ...)

vec_ptype2.Interval(x, y, ...)
}
\description{
vctrs methods for lubridate
}
\keyword{internal}
