\name{make_format}
\alias{make_format}
\title{Make format...}
\usage{make_format(order)}

\description{
  Make format
}

\details{
  Internal function used by parse functions.
}
\arguments{
  \item{order}{a character string that includes the letters y, m, and d in the preferred order of years, months, and dates}
}
