\name{mday-set}
\alias{mday-set}
\title{Internal function.}
\usage{\method{mday<-}{default} (x, value)}

\description{
  Internal function. Replaces the mdays element of a date with a
  specified value.
}
\seealso{\code{\link{mday}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's mdays component.}
}
