\name{minute-set}
\title{Internal function.}
\usage{minute<-(x, value)}

\description{
  Internal function. Replaces the minutes element of a date with a
  specified value.
}
\alias{minute<-.default}
\alias{minute<-.chron}
\alias{minute<-.zoo}
\alias{minute<-.its}
\alias{minute<-.ti}
\alias{minute<-.timeDate}
\alias{minute<-.jul}
\alias{minute<-.timeSeries}
\alias{minute<-.fts}
\alias{minute<-.irts}
\seealso{\code{\link{minute}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's minutes component.}
}
