\name{month-set}
\alias{month-set}
\title{Internal function.}
\usage{\method{month<-}{default} (x, value)}

\description{
  Internal function. Replaces the months element of a date with a
  specified value.
}
\seealso{\code{\link{month}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's months component.}
}
