\name{new_epoch}
\alias{new_epoch}
\title{Create an epoch object...}
\usage{new_epoch(weekday, number)}

\description{
  Create an epoch object
}

\details{
  new_epoch creates an epoch object with the specified
  values. Implementation for epochs will be implemented in
  future versions of lubridate.
}
\value{an epoch object}
\seealso{\code{\link{saturdays}}}
\arguments{
  \item{weekday}{name of weekday to use as epoch}
  \item{number}{number of epochs to include}
}
