\name{parse_unit_spec}
\alias{parse_unit_spec}
\title{Internal function.}
\usage{parse_unit_spec(unitspec)}

\description{
  Internal function. Parse date time unit specification
}

\details{
  Parse the time unit specification used by \code{\link{cut.Date}} into
  something useful
}
\keyword{internal}
