% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/periods.r
\name{period_to_seconds}
\alias{period_to_seconds}
\title{Convert a period to the number of seconds it appears to represent}
\usage{
period_to_seconds(x)
}
\arguments{
\item{x}{A period object}
}
\description{
Convert a period to the number of seconds it appears to represent
}

