% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epochs.r
\name{quick_epochs}
\alias{fridays}
\alias{mondays}
\alias{quick_epochs}
\alias{saturdays}
\alias{sundays}
\alias{thursdays}
\alias{tuesdays}
\alias{wednesdays}
\title{Quickly create common epoch objects}
\usage{
sundays(x = 1)

mondays(x = 1)

tuesdays(x = 1)

wednesdays(x = 1)

thursdays(x = 1)

fridays(x = 1)

saturdays(x = 1)
}
\arguments{
\item{x}{number of epochs to be included}
}
\description{
Creates an epoch object that uses the functions name as the reference epoch
event. Implementation for epochs will be implemented in future versions of
lubridate.
}
\keyword{internal}

