\name{reclass_date}
\title{Convenience method to reclass dates post-modification.}
\usage{reclass_date(new, orig)}

\description{
  Convenience method to reclass dates post-modification.
}
\keyword{internal}
\alias{reclass_date}
\alias{reclass_date.POSIXlt}
\alias{reclass_date.POSIXct}
\alias{reclass_date.chron}
\alias{reclass_date.timeDate}
\alias{reclass_date.its}
\alias{reclass_date.ti}
\alias{reclass_date.Date}
