\name{reclass_timespan}
\alias{reclass_timespan}
\alias{reclass_timespan-methods}
\title{Convenience method to reclass timespans post-modification.}
\usage{
  reclass_timespan(new, orig)

  reclass_timespan()
}
\description{
  Convenience method to reclass timespans
  post-modification.
}
\keyword{internal}

