\name{recognize}
\alias{recognize}
\title{Internal function.}
\usage{recognize(x)}

\description{
  Internal function. Is an object in a recognized date format?
}
\value{TRUE if x is a POSIXlt, POSIXct, yearmon, yearqtr, or Date object, FALSE otherwise.}
\keyword{internal}
\arguments{
  \item{x}{an R object}
}
