% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r, R/coercion.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{intersect}
\alias{union}
\alias{setdiff}
\alias{as.difftime}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:as.difftime]{as.difftime()}}, \code{\link[generics:intersect]{intersect()}}, \code{\link[generics:setdiff]{setdiff()}}, \code{\link[generics:union]{union()}}}
}}

