\name{second-set}
\alias{second-set}
\title{Internal function.}
\usage{\method{second<-}{default} (x, value)}

\description{
  Internal function. Replaces the seconds element of a date with a
  specified value.
}
\seealso{\code{\link{second}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's seconds component.}
}
