\name{shift}
\alias{shift}
\title{Shifts an interval up or down the timeline by a specified amount
note this may change the exact length of the interval}
\usage{
  shift(int, by)
}
\description{
  Shifts an interval up or down the timeline by a specified
  amount note this may change the exact length of the
  interval
}

