% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-of-day.r
\name{time_of_day}
\alias{time_of_day}
\title{Extract time of day in the current time zone}
\usage{
time_of_day(x)
}
\arguments{
\item{x}{a date-time object}
}
\value{
an object of class \linkS4class{Interval}
}
\description{
Extract time of day in the current time zone
}
\examples{
time_of_day(Sys.time())
}
