% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespans.r
\docType{class}
\name{Timespan-class}
\alias{Timespan-class}
\alias{*,Timespan,Timespan-method}
\alias{\%/\%,Timespan,Timespan-method}
\alias{\%/\%,difftime,Timespan-method}
\alias{\%/\%,Interval,Interval-method}
\alias{\%/\%,Period,Period-method}
\alias{\%/\%,Interval,Duration-method}
\alias{\%/\%,Duration,Interval-method}
\alias{\%/\%,Interval,Period-method}
\alias{\%/\%,Period,Interval-method}
\title{Timespan class}
\description{
Timespan is an S4 class with no slots. It is extended by the
\link[=Interval-class]{Interval}, \link[=Period-class]{Period}, and \link[=Duration-class]{Duration}
classes.
}
\keyword{internal}
