\name{tz-set}
\alias{tz-set}
\title{Internal function.}
\usage{\method{tz<-}{default} (x, value)}

\description{
  Internal function. Replaces the time zone element of a date with a
  specified value.
}
\seealso{\code{\link{tz}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's time zone component.}
}
