\name{wday-set}
\alias{wday-set}
\title{Internal function.}
\usage{\method{wday<-}{default} (x, value)}

\description{
  Internal function. Replaces the wdays element of a date with a
  specified value.
}
\seealso{\code{\link{wday}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's wdays component.}
}
