\name{week-set}
\alias{week-set}
\title{Internal function.}
\usage{\method{week<-}{default} (x, value)}

\description{
  Internal function. Replaces the weeks element of a date with a
  specified value.
}
\seealso{\code{\link{week}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's weeks component.}
}
