\name{yday-set}
\alias{yday-set}
\title{Internal function.}
\usage{\method{yday<-}{default} (x, value)}

\description{
  Internal function. Replaces the ydays element of a date with a
  specified value.
}
\seealso{\code{\link{yday}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's ydays component.}
}
