\name{year-set}
\alias{year-set}
\title{Internal function.}
\usage{\method{year<-}{default} (x, value)}

\description{
  Internal function. Replaces the years element of a date with a
  specified value.
}
\seealso{\code{\link{year}}}
\keyword{internal}
\keyword{methods}
\keyword{chron}
\keyword{manip}
\arguments{
  \item{x}{a date-time object}
  \item{value}{a number that will be substituted for the date's years component.}
}
