% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.r
\docType{methods}
\name{as_date}
\alias{as_date}
\alias{as_date,POSIXt-method}
\alias{as_date,character-method}
\alias{as_date,numeric-method}
\title{Change an object to a Date}
\usage{
as_date(x, ...)

\S4method{as_date}{POSIXt}(x, tzone = tz(x))

\S4method{as_date}{numeric}(x, epoch = as.Date("1970-01-01"))

\S4method{as_date}{character}(x, fmt = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{a vector of \code{\link{POSIXt}}, numeric or character objects}

\item{...}{further arguments to be passed to specific methods (see above).}

\item{tzone}{a time zone name (default: time zone of the POSIXt object
\code{x}). See \code{\link{olson_time_zones}}.}

\item{epoch}{a Date object, or something which can be coerced by
\code{as.Date(epoch, ...)} to such an object (default: the Unix epoch of 1
Jan 1970 as a \code{Date} object). Note that in this instance, \code{x} is
assumed to reflect the number of days since \code{eoch} at \code{"UTC"}.}

\item{fmt}{a character string (default: \code{\%Y-\%m-\%d}). See
\code{\link{strptime}} for alternative format specifications.}
}
\value{
a vector of \code{\link{Date}} objects corresponding to \code{x}.
}
\description{
as_date changes \code{\link{POSIXt}}, numeric and character objects to
\code{\link{Date}} objects.
}

