% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.r
\docType{methods}
\name{as_date}
\alias{as_date}
\alias{as_date,POSIXt-method}
\alias{as_date,numeric-method}
\alias{as_datetime}
\alias{as_datetime,ANY-method}
\alias{as_datetime,POSIXt-method}
\alias{as_datetime,numeric-method}
\title{Convert an object to a date or date-time}
\usage{
as_date(x, ...)

as_datetime(x, ...)

\S4method{as_date}{POSIXt}(x, tz = NULL)

\S4method{as_date}{numeric}(x, origin = lubridate::origin)

\S4method{as_datetime}{POSIXt}(x, tz = "UTC")

\S4method{as_datetime}{numeric}(x, origin = lubridate::origin, tz = "UTC")

\S4method{as_datetime}{ANY}(x, tz = "UTC")
}
\arguments{
\item{x}{a vector of \code{\link{POSIXt}}, numeric or character objects}

\item{...}{further arguments to be passed to specific methods (see above).}

\item{tz}{a time zone name (default: time zone of the POSIXt object
\code{x}). See \code{\link{OlsonNames}}.}

\item{origin}{a Date object, or something which can be coerced by
\code{as.Date(origin, ...)} to such an object (default: the Unix epoch of
"1970-01-01"). Note that in this instance, \code{x} is assumed to reflect
the number of days since \code{origin} at \code{"UTC"}.}
}
\value{
a vector of \code{\link{Date}} objects corresponding to \code{x}.
}
\description{
Convert an object to a date or date-time
}
\section{Compare to base R}{


These are drop in replacements for \code{as.Date} and \code{as.POSIXct},
with a few tweaks to make them work more intuitively.

\itemize{
  \item \code{as_date} ignores the timezone attribute, resulting in
     a more intuitive conversion (see examples)
  \item Both functions provide a default origin argument for numeric
     vectors.
  \item \code{as_datetime} defaults to using UTC.
}
}
\examples{
dt_utc <- ymd_hms("2010-08-03 00:50:50")
dt_europe <- ymd_hms("2010-08-03 00:50:50", tz="Europe/London")
c(as_date(dt_utc), as.Date(dt_utc))
c(as_date(dt_europe), as.Date(dt_europe))
## need not suply origin
as_date(10)
}

