% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coercion.r
\name{DateCoercion}
\alias{DateCoercion}
\title{Convert a variety of date-time classes to POSIXlt and POSIXct}
\description{
Convert a variety of date-time classes to POSIXlt and POSIXct
}
\keyword{internal}

