% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{Deprecated-lubridate}
\alias{Deprecated-lubridate}
\title{Deprecated functions in the lubridate package}
\arguments{
\item{x}{numeric value to be converted into duration}

\item{...}{arguments to be passed to the functions
(obscured to enforce the usage of new functions)}
}
\description{
Deprecated functions in the lubridate package
}
\keyword{internal}
