\name{DST}
\title{Internal function for Daylight Savings Time changes.}
\usage{DST(date1, date2)}

\description{
  Internal function for Daylight Savings Time changes.
}

\details{
  Determines how to handle time changes resulting from Daylight Savings
  time based on options("DST"). See \code{\link{DaylightSavingsTime}}.
}
\alias{DST}
\alias{DST.months}
\keyword{internal}
