% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ISO8601.r
\name{format_ISO8601_precision_check}
\alias{format_ISO8601_precision_check}
\title{Provide a format for ISO8601 dates and times with the requested precision.}
\usage{
format_ISO8601_precision_check(precision, max_precision, usetz = FALSE)
}
\arguments{
\item{precision}{The amount of precision to represent with
substrings of "ymdhms", as year, month, day, hour,
minute, and second. (e.g. "ymd" is days precision, "ymdhm" is minute precision.
When \code{NULL}, full precision for the object
is shown.}

\item{max_precision}{The maximum precision allowed to be output.}

\item{usetz}{Include the time zone in the formatting.  If
\code{usetz} is \code{TRUE}, the time zone is included. If
\code{usetz} is \code{"Z"}, the time is converted to "UTC" and
the time zone is indicated with "Z" ISO8601 notation.}
}
\description{
Provide a format for ISO8601 dates and times with the requested precision.
}
\keyword{internal}
