% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ISO8601.r
\name{format_ISO8601_precision_check}
\alias{format_ISO8601_precision_check}
\title{Provide a format for ISO8601 dates and times with the requested precision.}
\usage{
format_ISO8601_precision_check(precision, max_precision, usetz = FALSE)
}
\arguments{
\item{precision}{The amount of precision to represent with substrings of
"ymdhms", as "y"ear, "m"onth, "d"ay, "h"our, "m"inute, and
"s"econd. (e.g. "ymdhm" would show precision through minutes. When NULL,
\code{max_precision} is returned.  When \code{precision} is more precise than
\code{max_precision}, a warning is given and \code{max_precision} is returned.}

\item{max_precision}{The maximum precision allowed to be output.}

\item{usetz}{Include the timezone in the output format}
}
\description{
Provide a format for ISO8601 dates and times with the requested precision.
}
\keyword{internal}
