% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{ms}
\alias{hm}
\alias{hms}
\alias{ms}
\title{Create a period with the specified hours, minutes, and seconds}
\usage{
ms(..., quiet = FALSE)

hm(..., quiet = FALSE)

hms(..., quiet = FALSE)
}
\arguments{
\item{...}{a character vector of hour minute second triples}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
a vector of period objects
}
\description{
Transforms a character or numeric vector into a period object with the
specified number of hours, minutes, and seconds. hms() recognizes all
non-numeric characters except '-' as separators ('-' is used for negative
durations).  After hours, minutes and seconds have been parsed, the
remaining input is ingored.
}
\examples{
ms(c("09:10", "09:02", "1:10"))
ms("7 6")
ms("6,5")
hm(c("09:10", "09:02", "1:10"))
hm("7 6")
hm("6,5")

x <- c("09:10:01", "09:10:02", "09:10:03")
hms(x)

hms("7 6 5", "3:23:::2", "2 : 23 : 33", "Finished in 9 hours, 20 min and 4 seconds")
}
\seealso{
\code{\link{hm}, \link{ms}}
}
\keyword{period}

