% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r
\name{interval}
\alias{\%--\%}
\alias{interval}
\alias{is.interval}
\title{Create an \code{interval} object.}
\usage{
interval(start, end, tzone = attr(start, "tzone"))

start \%--\% end

is.interval(x)
}
\arguments{
\item{start}{a POSIXt or Date date-time object}

\item{end}{a POSIXt or Date date-time object}

\item{tzone}{a recognized timezone to display the interval in}

\item{x}{an R object}
}
\value{
an Interval object
}
\description{
interval creates an \code{\link{Interval-class}} object with the specified start and end
dates. If the start date occurs before the end date, the interval will be positive.
Otherwise, it will be negative.
}
\details{
Intervals are time spans bound by two real date-times.  Intervals can be
accurately converted to either period or duration objects using
\code{\link{as.period}}, \code{\link{as.duration}}. Since an interval is
anchored to a fixed history of time, both the exact number of seconds that passed
and the number of variable length time units that occurred during the interval can be
calculated.

\code{\%--\%} Creates an interval that covers the range spanned
by two dates. It replaces the
original behavior of lubridate, which created an interval by default whenever
two date-times were subtracted.
}
\examples{
interval(ymd(20090201), ymd(20090101))
# 2009-02-01 UTC--2009-01-01 UTC

date1 <- as.POSIXct("2009-03-08 01:59:59")
date2 <- as.POSIXct("2000-02-29 12:00:00")
interval(date2, date1)
# 2000-02-29 12:00:00 CST--2009-03-08 01:59:59 CST
interval(date1, date2)
# 2009-03-08 01:59:59 CST--2000-02-29 12:00:00 CST
span <- interval(ymd(20090101), ymd(20090201))
# 2009-01-01 UTC--2009-02-01 UTC

is.interval(period(months= 1, days = 15)) # FALSE
is.interval(interval(ymd(20090801), ymd(20090809))) # TRUE
}
\seealso{
\code{\link{Interval-class}}, \code{\link{as.interval}}
}

