% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespans.r
\name{is.timespan}
\alias{is.timespan}
\title{Is x a length of time?}
\usage{
is.timespan(x)
}
\arguments{
\item{x}{an R object}
}
\value{
TRUE if x is a period, interval, duration, or difftime object, FALSE otherwise.
}
\description{
Is x a length of time?
}
\examples{
is.timespan(as.Date("2009-08-03")) # FALSE
is.timespan(duration(second = 1)) # TRUE
}
\seealso{
\code{\link[=is.instant]{is.instant()}}, \code{\link[=is.duration]{is.duration()}}, \code{\link[=is.difftime]{is.difftime()}}, \code{\link[=is.period]{is.period()}}, \code{\link[=is.interval]{is.interval()}}
}
\keyword{chron}
\keyword{logic}
