\docType{data}
\name{lubridate_formats}
\alias{lubridate_formats}
\title{Lubridate format orders used in \code{stamp}}
\format{character vector of formats.}
\usage{
  lubridate_formats
}
\description{
  Lubridate format orders used in \code{stamp}
}
\seealso{
  \code{\link{parse_date_time}}, \code{\link{ymd}},
  \code{\link{ymd_hms}}
}
\keyword{chron}
\keyword{datasets}

