% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.r
\name{ms}
\alias{ms}
\title{Create a period with the specified number of minutes and seconds}
\usage{
ms(..., quiet = FALSE)
}
\arguments{
\item{...}{character or numeric vectors of minute second pairs}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
a vector of class \code{Period}
}
\description{
Transforms character or numeric vectors into a period object with the
specified number of minutes and seconds. ms() Arbitrary text can separate
minutes and seconds. Fractional separator is assumed to be ".". After
minutes and seconds have been parsed, all numeric input is ignored.
}
\examples{
ms(c("09:10", "09:02", "1:10"))
## [1] "9M 10S" "9M 2S"  "1M 10S"
ms("7 6")
## [1] "7M 6S"
ms("6,5")
## [1] "6M 5S"
}
\seealso{
\code{\link{hms}, \link{hm}}
}
\keyword{period}

