\name{new_epoch}
\alias{new_epoch}
\title{Create an epoch object}
\usage{
  new_epoch(weekday, number)
}
\arguments{
  \item{weekday}{name of weekday to use as epoch}

  \item{number}{number of epochs to include}
}
\value{
  an epoch object
}
\description{
  new_epoch creates an epoch object with the specified
  values. Implementation for epochs will be implemented in
  future versions of lubridate.
}
\seealso{
  \code{\link{saturdays}}
}

