% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.r
\name{now}
\alias{now}
\title{The current time}
\usage{
now(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like
the current time in. tzone defaults to your computer's system timezone.
You can retrieve the current time in the Universal Coordinated Time (UTC)
with now("UTC").}
}
\value{
the current date and time as a POSIXct object
}
\description{
The current time
}
\examples{
now()
now("GMT")
now("")
now() == now() # would be true if computer processed both at the same instant
now() < now() # TRUE
now() > now() # FALSE
}
\seealso{
\code{\link{here}}
}
\keyword{chron}
\keyword{utilities}

