\name{origin}
\alias{origin}
\title{1970-01-01 GMT...}

\description{
  1970-01-01 GMT
}

\details{
  Origin is the date-time for 1970-01-01 GMT in POSIXct
  format. This date-time is the origin for the numbering
  system used by POSIXct, POSIXlt, chron, and Date classes.
}
\keyword{data}
\keyword{chron}
\examples{origin
# "1970-01-01 GMT"}
