\docType{data}
\name{origin}
\alias{origin}
\title{1970-01-01 GMT}
\format{POSIXct[1:1], format: "1970-01-01"}
\usage{
  origin
}
\description{
  Origin is the date-time for 1970-01-01 GMT in POSIXct
  format. This date-time is the origin for the numbering
  system used by POSIXct, POSIXlt, chron, and Date classes.
}
\examples{
origin
# "1970-01-01 GMT"
}
\keyword{chron}
\keyword{data}
\keyword{dataset}

