% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/instants.r
\docType{data}
\name{origin}
\alias{origin}
\title{1970-01-01 UTC}
\format{\preformatted{ POSIXt[1:1], format: "1970-01-01"
}}
\usage{
origin
}
\description{
Origin is the date-time for 1970-01-01 UTC in POSIXct format. This date-time
is the origin for the numbering system used by POSIXct, POSIXlt, chron, and
Date classes.
}
\examples{
origin
# "1970-01-01 GMT"
}
\keyword{chron}
\keyword{data}

