% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.r
\docType{data}
\name{origin}
\alias{origin}
\title{1970-01-01 UTC}
\format{An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.}
\usage{
origin
}
\description{
Origin is the date-time for 1970-01-01 UTC in POSIXct format. This date-time
is the origin for the numbering system used by POSIXct, POSIXlt, chron, and
Date classes.
}
\examples{
origin
}
\keyword{chron}
\keyword{data}

