\name{parse_date}
\alias{parse_date}
\title{Deprecated internal function, use \code{\link{parseDateTime}} instead.}
\usage{
  parse_date(x, formats, quiet = FALSE, seps = NULL,
    tz = "UTC")
}
\arguments{
  \item{x}{-}

  \item{formats}{-}

  \item{quiet}{-}

  \item{seps}{-}

  \item{tz}{-}
}
\description{
  Deprecated internal function, use
  \code{\link{parseDateTime}} instead.
}

