\name{period_to_seconds}
\alias{period_to_seconds}
\title{Convert a period to the number of units it appears to represent}
\usage{
  period_to_seconds(x)

  period_to_seconds(x)
}
\arguments{
  \item{x}{A period object}

  \item{x}{A period object}
}
\description{
  Convert a period to the number of units it appears to
  represent

  Convert a period to the number of units it appears to
  represent
}

