% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-quarter.r
\name{quarter}
\alias{quarter}
\title{Get the calendar or fiscal quarter of a date-time.}
\usage{
quarter(x, with_year = FALSE, fiscal_start = 1)
}
\arguments{
\item{x}{a date-time object of class POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, fts or anything else that can be converted
with as.POSIXlt}

\item{with_year}{logical indicating whether or not to include the quarter's year.}

\item{fiscal_start}{numeric indicating the month in which the fiscal year starts, e.g. 11 for November}
}
\value{
numeric the fiscal quarter that the date-time occurs in
}
\description{
Calendar quarters are a way of dividing the year into fourths. The first quarter (Q1)
comprises January, February and March; the second quarter (Q2) comprises April, May,
June; the third quarter (Q3) comprises July, August, September; the fourth quarter (Q4)
October, November, December. Fiscal quarters often do not align with calendar quarters,
but are shifted such that the beginning of the fiscal year is a set number of months prior or
subsequent to January.
}
\examples{
x <- ymd(c("2012-03-26", "2012-05-04", "2012-09-23", "2012-12-31"))
quarter(x)
# 1 2 3 4
quarter(x, with_year = TRUE)
# 2012.1 2012.2 2012.3 2012.4
quarter(x, with_year = TRUE, fiscal_start = 11)
# 2012.2 2012.3 2012.4 2013.1
}
\keyword{chron}
\keyword{manip}
\keyword{methods}
\keyword{utilities}

