% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coercion.r
\name{reclass_date}
\alias{reclass_date}
\title{Convenience method to reclass dates post-modification.}
\usage{
reclass_date(new, orig)
}
\description{
Convenience method to reclass dates post-modification.
}
\keyword{internal}

