% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coercion.r
\name{reclass_timespan}
\alias{reclass_timespan}
\title{Convenience method to reclass timespans post-modification.}
\usage{
reclass_timespan(new, orig)
}
\description{
Convenience method to reclass timespans post-modification.
}
\keyword{internal}

