% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/coercion.r
\name{reclass_timespan}
\alias{reclass_timespan}
\alias{reclass_timespan,ANY,Duration-method}
\alias{reclass_timespan,ANY,Interval-method}
\alias{reclass_timespan,ANY,Period-method}
\alias{reclass_timespan,ANY,difftime-method}
\title{Convenience method to reclass timespans post-modification.}
\usage{
reclass_timespan(new, orig)

reclass_timespan(new, orig)
}
\description{
Convenience method to reclass timespans post-modification.

#'
}
\keyword{internal}

