\name{reclass_timespan}
\alias{reclass_timespan}
\alias{reclass_timespan,ANY,Duration-method}
\alias{reclass_timespan,ANY,Interval-method}
\alias{reclass_timespan,ANY,Period-method}
\alias{reclass_timespan,ANY,difftime-method}
\title{Convenience method to reclass timespans post-modification.}
\usage{
  reclass_timespan(new, orig)
}
\description{
  Convenience method to reclass timespans
  post-modification.
}
\keyword{internal}

