% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.r, R/coercion.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{intersect}
\alias{union}
\alias{setdiff}
\alias{as.difftime}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:coercion-time-difference]{as.difftime}}, \code{\link[generics:setops]{intersect}}, \code{\link[generics:setops]{setdiff}}, \code{\link[generics:setops]{union}}}
}}

