% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.r
\name{round_date}
\alias{ceiling_date}
\alias{floor_date}
\alias{round_date}
\title{Round, floor and ceiling methods for date-time objects.}
\usage{
round_date(x, unit = "second")

floor_date(x, unit = "seconds")

ceiling_date(x, unit = c("second", "minute", "hour", "day", "week", "month",
  "bimonth", "quarter", "halfyear", "year"), change_on_boundary = FALSE)
}
\arguments{
\item{x}{a vector of date-time objects}

\item{unit}{a character string specifying the time unit or a multiple of a
unit to be rounded to. Valid base units are second, minute, hour, day,
week, month, bimonth, quarter, halfyear, or year. Unique abriviation or
plurals are also supported (min, mins, secs etc). Rounding to multiple of
units (except weeks) is suported from \code{v1.6.0}.}

\item{change_on_boundary}{logical. If FALSE (the default) \code{ceiling_date}
don't alter date-times on the boundary of the \code{unit}. The boundary is
unit dependent - for second, minute, hour and day the boundary is `0` units
of next smaller unit. For week, month and year the boundary is on the first
day within that unit. For example for the boundary date "2000-01-01"
\code{ceiling_date(ymd("2000-01-01"), "month")} is \code{"2000-01-01"}
while `ceiling_date(ymd("2000-01-01"), "month", TRUE)` is "2000-02-01".}
}
\value{
x with the appropriate units floored
}
\description{
Rounding to the nearest unit or multiple of a unit are supported. All
meaningfull specifications in english are supported - secs, min, mins, 2
minutes, 3 years etc.
}
\details{
\code{round_date} takes a date-time object and rounds it to the nearest value
of the specified time unit. For rounding date-ties which is exactly halfway
between two consecutive units, the convention is to round up. Note that this
is in line with the behavior of R's base \link[base]{round.POSIXt} function
but does not follow the convention of the base \link[base]{round} function
which "goes to the even digit" per IEC 60559.

\code{floor_date} takes a date-time object and rounds it down to the nearest
value of the specified time unit.

\code{ceiling_date} takes a date-time object and rounds it up to the nearest
value of the specified time unit.

By convention the boundary for a month is the 00 second of the month. Thus
\code{ceiling_date(ymd("2000-03-01"), "month")} is equivalent to
\code{ceiling_date(ymd_hms("2000-03-01 00:00:00"), "month")} and returns
"2000-03-01 UTC". Use \code{change_on_boundary} to alter this behavior.
}
\examples{
x <- as.POSIXct("2009-08-03 12:01:59.23")
round_date(x, "second")
round_date(x, "minute")
round_date(x, "5 mins")
round_date(x, "hour")
round_date(x, "2 hours")
round_date(x, "day")
round_date(x, "week")
round_date(x, "month")
round_date(x, "bimonth")
round_date(x, "quarter") == round_date(x, "3 months")
round_date(x, "halfyear")
round_date(x, "year")

x <- as.POSIXct("2009-08-03 12:01:59.23")
floor_date(x, "second")
floor_date(x, "minute")
floor_date(x, "hour")
floor_date(x, "day")
floor_date(x, "week")
floor_date(x, "month")
floor_date(x, "bimonth")
floor_date(x, "quarter")
floor_date(x, "halfyear")
floor_date(x, "year")

x <- as.POSIXct("2009-08-03 12:01:59.23")
ceiling_date(x, "second")
ceiling_date(x, "minute")
ceiling_date(x, "5 mins")
ceiling_date(x, "hour")
ceiling_date(x, "day")
ceiling_date(x, "week")
ceiling_date(x, "month")
ceiling_date(x, "bimonth") == ceiling_date(x, "2 months")
ceiling_date(x, "quarter")
ceiling_date(x, "halfyear")
ceiling_date(x, "year")
x <- ymd("2000-01-01")
ceiling_date(x, "month")
ceiling_date(x, "month", change_on_boundary = TRUE)
}
\seealso{
\link[base]{round}
}
\keyword{chron}
\keyword{manip}

