\name{subtract_dates}
\alias{subtract_dates}
\title{Subtracting date-time objects}
\usage{
  subtract_dates(e1, e2)
}
\arguments{
  \item{e1}{a numeric or date-time object}

  \item{e2}{a numeric or date-time object}
}
\description{
  subtract_dates subtracts two objects, both of which can
  be date-time objects. Lubridate is built so that you
  don't need to use subtract_dates but can use regular math
  expressions instead.
}

