\name{sundays}
\title{Quickly create common epoch objects...}
\usage{sundays(x=1)}

\description{
  Quickly create common epoch objects
}

\details{
  Creates an epoch object that uses the functions name as the reference
  epoch event. Implementation for epochs will be implemented in future
  versions of lubridate.
}
\alias{sundays}
\alias{mondays}
\alias{tuesdays}
\alias{wednesdays}
\alias{thursdays}
\alias{fridays}
\alias{saturdays}
