\name{sundays}
\alias{fridays}
\alias{mondays}
\alias{saturdays}
\alias{sundays}
\alias{thursdays}
\alias{tuesdays}
\alias{wednesdays}
\title{Quickly create common epoch objects}
\usage{
  sundays(x = 1)
}
\arguments{
  \item{x}{number of epochs to be included}
}
\description{
  Creates an epoch object that uses the functions name as
  the reference epoch event. Implementation for epochs will
  be implemented in future versions of lubridate.
}

