% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespans.r
\docType{class}
\name{Timespan-class}
\alias{*,Timespan,Timespan-method}
\alias{Timespan-class}
\alias{\%/\%,Timespan,Timespan-method}
\alias{\%/\%,difftime,Timespan-method}
\title{Timespan class}
\description{
Timespan is an S4 class with no slots. It is extended by the
\code{\link{Interval-class}}, \code{\link{Period-class}}, and \code{\link{Duration-class}}
classes.
}

