% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-tz.r
\name{tz}
\alias{tz}
\alias{tz<-}
\title{Get/set time zone component of a date-time}
\usage{
tz(x)

tz(x) <- value
}
\arguments{
\item{x}{A date-time vector, usually of class POSIXct or POSIXlt.}

\item{value}{timezone value to be assigned to \code{x}'s \code{tzone} attribute}
}
\value{
A character vector of length 1, giving the time zone of the vector.
An empty string (\code{""}) represents the current/default timezone.

For backward compatibility, the time zone of a date, \code{NA}, or
character vector is \code{"UTC"}.
}
\description{
Conveniently get and set the time zone of date-time, with convenient
fallback behaviour for dates. Note that modifying the time does not change
the instant in time represented by the vector, just its printed
representation. Use \code{\link[=with_tz]{with_tz()}} if need the same time (i.e. a different
instant) in another time zone.
}
\section{Valid time zones}{

Time zones are stored in system specific database, so are not guaranteed
to be the same on every system (however, they are usually pretty similar
unless your sytsem is very out of date). You can see a complete list with
\code{\link[=OlsonNames]{OlsonNames()}}
}

\examples{
x <- ymd("2012-03-26", tz = "UTC")
tz(x)

tz(x) <- "GMT"
x
}
\seealso{
See \link{DateTimeClasses} for a description of the underlying
\code{tzone} attribute..
}
\keyword{chron}
\keyword{manip}
\keyword{methods}
\keyword{utilities}
